web/web_requirements.txt
Flask
requests
lxml
pymongo
python-dateutil
